assignment1/requirements.txt
Flask
flask_restful
io
spacy
streamlit
annotated_text
matplotlib.pyplot
